Fix some warnings.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@48475 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ E16/epplets/epplets/E-Xss.c
@@ -281,7 +281,7 @@ static void
 config_cb(void *data)
 {
    char                buff[128];
-   unsigned long       h, y, i;
+   unsigned int        h, y, i;
 
    if (config_win)
      {
